Remove references to MediatR behaviors
