--- src/ConductorSharp.Engine/Behaviors/TaskExecutionTrackingBehavior.cs
@@ -1,66 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
-using ConductorSharp.Engine.Interface;
-using ConductorSharp.Engine.Util;
-
-namespace ConductorSharp.Engine.Behaviors
-{
-    public class TaskExecutionTrackingBehavior<TRequest, TResponse> : INgWorkerMiddleware<TRequest, TResponse>
-        where TRequest : class, ITaskInput<TResponse>, new()
-    {
-        private readonly ConductorSharpExecutionContext _executionContext;
-        private readonly IEnumerable<ITaskExecutionService> _taskExecutionServices;
-
-        public TaskExecutionTrackingBehavior(
-            ConductorSharpExecutionContext executionContext,
-            IEnumerable<ITaskExecutionService> taskExecutionServices
-        )
-        {
-            _executionContext = executionContext;
-            _taskExecutionServices = taskExecutionServices;
-        }
-
-        public async Task<TResponse> Handle(
-            TRequest request,
-            Func<TRequest, CancellationToken, Task<TResponse>> next,
-            CancellationToken cancellationToken
-        )
-        {
-            var trackedTask = new RunningTask
-            {
-                TaskId = _executionContext.TaskId,
-                TaskName = _executionContext.TaskName,
-                StartedAt = DateTimeOffset.UtcNow
-            };
-
-            foreach (var taskExecutionService in _taskExecutionServices)
-            {
-                await taskExecutionService.OnPolled(trackedTask);
-            }
-
-            try
-            {
-                var response = await next(request, cancellationToken);
-
-                foreach (var taskExecutionService in _taskExecutionServices)
-                {
-                    await taskExecutionService.OnCompleted(trackedTask);
-                }
-
-                return response;
-            }
-            catch (Exception)
-            {
-                foreach (var taskExecutionService in _taskExecutionServices)
-                {
-                    await taskExecutionService.OnFailed(trackedTask);
-                }
-
-                throw;
-            }
-        }
-    }
-}
