--- src/ConductorSharp.Engine/Exceptions/BaseWorkerException.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace ConductorSharp.Engine.Exceptions
-{
-    public class BaseWorkerException : Exception
-    {
-        public BaseWorkerException(string message, Exception innerException) : base(message, innerException) { }
-    }
-}
